feat: add release script and changelogen dependency
This commit introduces a new script for releasing the package. The
"release" script uses the "changelogen" package to generate a changelog
and then publishes the package to npm. It also pushes the tags to the
repository. The "changelogen" package has been added as a devDependency
in the package.json file. The lock file has been updated accordingly.