feat: replace estree-walker with zimmerframe
This commit replaces the estree-walker package with zimmerframe in
src/preprocessor.ts. The walk function from zimmerframe is used to
traverse the AST. The commit also refactors the budouxPreprocess
function to use a state array for storing nodes and their parsed
values, which are then used to overwrite the original nodes. This
change improves the efficiency of the preprocessor.